pyproject.toml
[tool.poetry]
name = "deep-subtitle-translate"
version = "0.1.0"
description = ""
authors = ["Leonardo Cardoso <lcard00@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
